# docs/swagger.yaml
basePath: /
definitions:
  handler.getAllListsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/restapi.TodoList'
        type: array
    type: object
  handler.myError:
    properties:
      message:
        type: string
    type: object
  handler.signInInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  restapi.ListItem:
    properties:
      id:
        type: integer
      itemId:
        type: integer
      listId:
        type: integer
    type: object
  restapi.TodoList:
    properties:
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
  restapi.User:
    properties:
      name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - name
    - password
    - username
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for TodoList Application
  title: Todo App API
  version: "1.0"
paths:
  /api/lists:
    get:
      consumes:
      - application/json
      description: get all lists
      operationId: get-all-lists
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.getAllListsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.myError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.myError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.myError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.myError'
      security:
      - ApiKeyAuth: []
      summary: Get All Lists
      tags:
      - lists
    post:
      consumes:
      - application/json
      description: create todo list
      operationId: create-list
      parameters:
      - description: list info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/restapi.TodoList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.myError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.myError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.myError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.myError'
      security:
      - ApiKeyAuth: []
      summary: Create todo list
      tags:
      - lists
  /api/lists/:id:
    get:
      consumes:
      - application/json
      description: get list by id
      operationId: get-list-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restapi.ListItem'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.myError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.myError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.myError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.myError'
      security:
      - ApiKeyAuth: []
      summary: Get List By Id
      tags:
      - lists
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login
      operationId: login
      parameters:
      - description: credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.myError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.myError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.myError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.myError'
      summary: SignIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create account
      operationId: create-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/restapi.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.myError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.myError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.myError'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.myError'
      summary: SignUp
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"